+++ src/chromatix/utils/__init__.py
@@ -2,3 +2,4 @@ from .utils import *
 from .data import *
 from .shapes import *
 from .initializers import *
+from .grids import *
