+++ src/chromatix/utils/initializers.py
@@ -2,8 +2,8 @@ import jax.numpy as jnp
 from einops import rearrange
 from chex import Array
 from typing import Sequence, Tuple
-from .utils import create_grid, grid_spatial_to_pupil
-from scipy.special import comb
+from .grids import create_grid, grid_spatial_to_pupil
+from scipy.special import comb #type: ignore
 import math
 
 __all__ = [
